Add an "incompatible" constraint value
The bazel docs [1] currently recommend users to create an
`incompatible` constraint in their own code base. This constraint is
useful in expressing more advanced target compatibilities.

Currently every project has to create their own definition for this
constraint. That is unnecessary proliferation of something that can be
easily centralized here.

This patch makes it so users will be able to use the following pattern
in their BUILD files:

    cc_library(
        name = "unixish_lib",
        srcs = "unixish_lib.cc",
        target_compatible_with = select({
            "@platforms//os:osx": [],
            "@platforms//os:linux": [],
            "//conditions:default": ["@platforms//:incompatible"],
        ],
    )

[1]: https://docs.bazel.build/versions/master/platforms.html#skipping-incompatible-targets