Fix platform_data
These fixes are necessary for platform_data to work. Tested by building
http://pwrev.dev/189772.

I was long confused by the need for the [0] index. It turns out this
arises because "there's no easy distinction between patch and split
transitions, and so all transitions require the index, even if it's not
a split."